Remove case insensitivity from .goob
I recieved a bug report indicating that TARS had somehow sent a message into a channel that it wasn't a member of. The only possible explanation I can think of is that the vowel-replacement modified the ACTION call. This commit makes the check case-sensitive so that it can't modify the ACTION, which is always capitalised.