Invert GraphQL iteration direction
I've been having some problems iterating between GraphQL response pages.
When setting the order as DESC, I expect the list to be inherently
ordered in that direction; instead, it appears that the list is simply
iterated backwards. Opening up the pagination cursors in Base64 appears
to indicate that the sort order is not used to generate the cursor,
which could imply that pagination simply ignores sort order. However,
this is my first time consuming a GraphQL API, so it is more likely that
this is simply a result of my inexperience.

Switching the sort order to ASC and simply iterating backwards seems to
work, though truthfully, I've no idea why endCursor is still the correct
cursor to use and not startCursor.