Use case-insensitive search for SCP numbers
A search query that is a number goes through the following matches:

1. Match articles whose title contains the number.
2. Match articles whose SCP number is the number (for matching strings
like "SCP-nnnn")
3. Match articles whose SCP number contains the number (for matching
string like "nnnn" against SCP-nnnn).

Previously, the stored SCP title was normalised to lowercase. Using the
Crom API, the methods that wrap it store the value directly without
case-normalising it. This is preferable as there's no guarantee that an
SCP number is case insensitive (it may be required to retrieve and
display it in the intended case at some point). However, the match was
being performed case-sensitively.

This commit uses the case-insensitive regex match for matching the SCP
number, as that's the only case-insensitive SQL method I had to hand. It
will probably be a bit too permissive, but I really, really don't care.