Extract out commands registry to a helper
The command registry consisted of the registry declaration itself (a big
object) and then a whole bunch of messy stuff to parse the object. This
was poorly-designed and honestly a miracle it worked.

This commit extracts the functionality for parsing out to a new helper
that only contains the command registry class.