Split command instantiation and parsing
The documentation requires the argparse usage string, but the usage
string can only be created from an instantiated command. Instantiation
required the command message in order to parse it. During documentation
generation, there are no command messages, and therefore no way to
instantiate a command.

This commit breaks out parsing into its own method on the base command,
which needs to be called after instantiation but before execution.
Instantiation no longer takes an argument, which worries me in that
maybe it's not necessary and never has been and maybe I need to refactor
everything again, but at least it means I can get the argparse message
during documentation generation.