Convert from Webpack to Snowpack
Moving away from the Webpack dev server to Snowpack's built-in dev
server and unbundled dev experience ideology, rebuild times for the dev
server have been reduced from about 20 seconds to about 4 seconds, with
the webpage still being visible in the interim.

However, Snowpack has trouble resolving imports in CSS files in npm
packages, resulting in it being unable to find the font files for the
Fontsource package. I have created a pre-build script that copies the
font files to the location that Snowpack erroneously expects to find
them per snowpackjs/snowpack#1573. This location is fortunately in
build/ which is already ignored.