Construct document as nested sections
In researching how to highlight the currently-visible part of the page
in the sidebar, I have discovered the Intersection Observer API, which
will let me see which parts of the page are currently visible. The
example I saw highlighted in the sidebar not only which subsection was
currently visible but also which parent section - something that I'd
been planning on doing already. The Intersection Observer API should let
me do that with very little code.

However, in order to highlight both the child and parent sections, I
would need to know when each of them are on the page. All the data used
to examine the flow of the document (and hence the structure of the
sidebar) was stored in the heading elements. The headings are only on
screen for a brief moment, but I would need the IO API to be able to
tell when their respective sections are on screen. I thought that the
easiest way for me to do this would be to move all the data to the
sections.

This would also enable me to move away from exposing the origin of the
header context as a CSS class that is the same as the original header's
tag, which was not semantic. I did this by adding a data-role attribute,
which stores the context of the section.

For most of the page this was fairly easy, but I wanted each layer of
heading to be the root of its own tree of <section>s, and for the part
of the page that is generated directly from Markdown, this would not be
easy. I ended up creating a regex replace function that strips a layer
of headings from the Markdown and wraps it in a section. It is then
called in a loop to iterate through all headings in the document until
they are all wrapped in a section. This supersedes the old Markdown
preprocessing step where a heading's ID is generated from its content:
the algorithm is still present but now the ID is attached to the
section. I had to invoke a Python-Markdown extension to process the
Markdown that was now inside HTML tags.

This commits also adjusts the CSS for the new structure.