Remove controller checks where appropriate
I want to use the base command's permission system instead of manually
checking if the current user is a Controller as much as possible. The
system isn't a catch-all, though, because it can't handle cases where a
permission level is needed based on an argument's value rather than its
presence e.g. adding aliases to a nick that is not your own.

This commit effectively disables the existing rudimentary permissions
system and must not be merged until another commit implements the new
permissions system.