Create an argparse action that checks permissions
Permission checking on a per-argument basis (as opposed to per-command)
now works via a custom argparse action.

All actions were previously 'store' except for when type was bool in
which case it was 'store_true'. This dichotomy still exists except now
the choice changes the parent of the derived Action class rather than a
parameter passed to add_argument.

The base command does not come with a permission checker - one must be
passed to it during instantiation. This ensures that the base command is
context-agnostic. With specific permission checkers passed to it, it
should work in any context be that IRC, Discord or documentation
generation.