Check that self import works
Tars is supposed to be booted as a module using the -m flag on the
command line but actually all this does is add the module's path to
sys.path and then boot the module's __main__ file as a script. os.execl,
which is used to reboot the bot, is not aware of the changes made to
sys.path and restarts the process as it believes it should be - by
running __main__ as a script. This doesn't work because it doesn't keep
the changes to sys.path, which means tars can't import itself.

A self-import check has been added, if it fails, the package's path is
added to sys.path.