Do not raise CommandNotExistError in registry
The registry raised CommandNotExistError when attempting to access a
command by alias, which sounds right, but was specific to the
implementation of try_command in the parsemessages plugin. I want things
to be as generic as possible wherever possible, so all checks have been
removed and it will just raise AttributeError from the dict access if no
such alias exists.

This commit also renames get_command to get_command_by_alias to be more
consistent with (and distinct from) get_command_by_name; and commands is
imported from the tars module rather than importing tars.commands to
reduce line length, which was not truncated during blackening.