Switch from intermixed to regular args
In 638f3ee I switch from regular args to intermixed, erroneously
believing that was the only way I could allow positional arguments to
come before optional arguments. This is not the case: intermixed args
are for when the optionals are divided by the positionals, which I see
no reason to permit.

I had configured .seen really weirdly in order to allow the positional
to come after the optional; however, this was entirely unnecessary. So
now it's back to normal.