Add longstr as argument type
The longstr extends str. It doesn't do anything itself, but if the
custom parsing action sees it, it will concatenate the given arguments
into a single string joined by single spaces. It should be used with any
nargs value that would normally expose a list, but actually exposes a
single value.