Switch from Snowpack to Webpack
Webpack was taking really long for dev rebuilds, so I switched to
Snowpack, which promised much faster speeds. Snowpack delivered on that
promise, playing really nicely with Tailwind's huge dev CSS sizes, but
it couldn't handle relative imports in CSS files, which broke
Fontsource. I was able to hack together a solution that fixed it on
local, and thought that would last; however, it doesn't work in GitHub
Actions' environment and I have not the mental strength to try to work
out why.

A recent update to Tailwind brings a new experimental just-in-time
engine, which explicitly promised better performance with Webpack. I
finally decided to give it a go and the results are amazing - damn close
to instant dev rebuilds, and a full prod build in two seconds. Hopefully
this will work with Actions.