Fix prefix storage
I added a shortcut where a prefix of length 0 would not be stored to
the command object, resulting in the default value of None, allowing me
to safely check whether there was a prefix by looking for None. However,
while I intended this to be for prefixes of length greater than 0, it
was actually looking for prefixes of length greater than 1. Helen's
prefix is length one, so this was preventing any and all deferral.

Also removes a stray print statement.