Assign default nargs of None
The default nargs is assumed to be None by argparse, and I've been
assuming the same; however, the part of the parser that sets the default
default argument value only actually executes if the nargs is provided.
This was because it needs to check the value of the nargs to see which
default to set, which would error if the nargs were not provided. I have
replaced the check for nargs to simply explicitly set a default value of
None if it was not provided.