Do not set default nargs if type is bool
In the previous commit I added a default nargs value of None; however,
for boolean argument, nargs must either be 0 or not present. Because I
mostly define boolean arguments with undefined args, these defaulted to
None, which broke them. Excluding boolean-type arguments from this check
avoid that, though I had to move the boolean argument handler below this
check because the handler actually removes the nargs if provided (which
would've cause the problem all over again).